Refactor _types pt.6
