+++ specification/xpack/usage/types.ts
@@ -21,7 +21,7 @@ import { Phases } from '@ilm/_types/Phases'
 import { Statistics } from '@slm/_types/SnapshotLifecyclePolicyMetadata'
 import { Dictionary } from '@spec_utils/Dictionary'
 import { ByteSize, EmptyObject, Field, Name } from '@_types/common'
-import { Job, JobStatistics } from '@_types/ml/Job'
+import { Job, JobStatistics } from '@ml/_types/Job'
 import { integer, long, uint, ulong } from '@_types/Numeric'
 
 export class Base {
